Move old Java-based plugin API code to internal scope
