+++ plugin-infra/go-plugin-access/src/main/java/com/thoughtworks/go/plugin/access/secrets/VersionedSecretsExtension.java
@@ -21,7 +21,9 @@ import com.thoughtworks.go.plugin.domain.common.Image;
 import com.thoughtworks.go.plugin.domain.common.PluginConfiguration;
 import com.thoughtworks.go.plugin.domain.secrets.Secret;
 
-import java.util.*;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 public interface VersionedSecretsExtension {
     Image getIcon(String pluginId);
