+++ plugin-infra/go-plugin-access/src/test/java/com/thoughtworks/go/plugin/access/elastic/v4/ElasticAgentExtensionConverterV4Test.java
@@ -30,7 +30,8 @@ import static java.lang.String.format;
 import static net.javacrumbs.jsonunit.fluent.JsonFluentAssert.assertThatJson;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class ElasticAgentExtensionConverterV4Test {
     private JobIdentifier jobIdentifier;
