+++ plugin-infra/go-plugin-api-internal/src/main/java/com/thoughtworks/go/plugin/api/material/packagerepository/PackageMaterialProperty.java
@@ -15,24 +15,29 @@
  */
 package com.thoughtworks.go.plugin.api.material.packagerepository;
 
-import java.util.Collections;
-import java.util.List;
-
-import com.thoughtworks.go.plugin.api.config.Configuration;
 import com.thoughtworks.go.plugin.api.config.Property;
 
-/**
- * Represents {@link Configuration} specific to repository
- */
+public class PackageMaterialProperty extends Property implements Comparable {
+    public PackageMaterialProperty(String key) {
+        super(key);
+        updateDefaults();
+    }
 
-@Deprecated
-//Will be moved to internal scope
-public class RepositoryConfiguration extends Configuration {
+    private void updateDefaults() {
+        with(REQUIRED, true);
+        with(PART_OF_IDENTITY, true);
+        with(SECURE, false);
+        with(DISPLAY_NAME, "");
+        with(DISPLAY_ORDER, 0);
+    }
+
+    public PackageMaterialProperty(String key, String value) {
+        super(key, value);
+        updateDefaults();
+    }
 
     @Override
-    public List<? extends Property> list() {
-        List<PackageMaterialProperty> list = (List<PackageMaterialProperty>) super.list();
-        Collections.sort(list);
-        return list;
+    public int compareTo(Object o) {
+        return this.getOption(DISPLAY_ORDER) - ((PackageMaterialProperty) o).getOption(DISPLAY_ORDER);
     }
 }
