+++ plugin-infra/go-plugin-api-internal/src/main/java/com/thoughtworks/go/plugin/api/response/execution/ExecutionResult.java
@@ -20,8 +20,6 @@ import com.thoughtworks.go.plugin.api.response.Result;
 /**
  * Used to specify the result of an execution of a part of a plugin.
  */
-@Deprecated
-//Will be moved to internal scope
 public class ExecutionResult extends Result {
 
     /**
