+++ plugin-infra/go-plugin-api-internal/src/main/java/com/thoughtworks/go/plugin/api/response/validation/ValidationError.java
@@ -19,8 +19,6 @@ package com.thoughtworks.go.plugin.api.response.validation;
 /**
  * Represents error reported when validation fails
  */
-@Deprecated
-//Will be moved to internal scope
 public class ValidationError {
 
     private static final String EMPTY_KEY = "";
