+++ plugin-infra/go-plugin-api-internal/src/main/java/com/thoughtworks/go/plugin/api/response/validation/ValidationResult.java
@@ -22,8 +22,6 @@ import java.util.List;
 /**
  * Represents result of validation
  */
-@Deprecated
-//Will be moved to internal scope
 public class ValidationResult {
 
     private List<ValidationError> errors = new ArrayList<>();
