+++ plugin-infra/go-plugin-api-internal/src/main/java/com/thoughtworks/go/plugin/api/task/Console.java
@@ -21,8 +21,6 @@ import java.util.Map;
 /**
  * Used to write information of a task run, out to the build log.
  */
-@Deprecated
-//Will be moved to internal scope
 public interface Console {
     /**
      * Print a line out to the build log.
