+++ plugin-infra/go-plugin-api-internal/src/main/java/com/thoughtworks/go/plugin/api/task/TaskConfig.java
@@ -28,8 +28,6 @@ import java.util.List;
  * This class, at other times, is used to hold information about the value provided by the
  * user, for the configuration.
  */
-@Deprecated
-//Will be moved to internal scope
 public class TaskConfig extends Configuration {
     /**
      * Adds a property to the configuration of this task.
