+++ plugin-infra/go-plugin-api-internal/src/main/java/com/thoughtworks/go/plugin/api/task/TaskExecutionContext.java
@@ -19,8 +19,6 @@ package com.thoughtworks.go.plugin.api.task;
  * Used to provide the {@link TaskExecutor} with context about the
  * environment of the task run.
  */
-@Deprecated
-//Will be moved to internal scope
 public interface TaskExecutionContext {
     /**
      * The environment variables that this task should run with.
