+++ plugin-infra/go-plugin-api-internal/src/main/java/com/thoughtworks/go/plugin/api/task/TaskExecutor.java
@@ -21,8 +21,6 @@ import com.thoughtworks.go.plugin.api.response.execution.ExecutionResult;
  * The implementation of this TaskExecutor interface will be the one which does the actual work of the task.
  *
  */
-@Deprecated
-//Will be moved to internal scope
 public interface TaskExecutor {
     /**
      * The plugin infrastructure will call the execute() method, on the agent side, when the task needs to be run.
