+++ server/src/main/java/com/thoughtworks/go/server/service/RuleAwarePluginProfileService.java
@@ -17,10 +17,10 @@ package com.thoughtworks.go.server.service;
 
 
 import com.thoughtworks.go.config.ConfigTag;
-import com.thoughtworks.go.config.rules.RuleAwarePluginProfile;
-import com.thoughtworks.go.config.rules.RuleAwarePluginProfiles;
 import com.thoughtworks.go.config.exceptions.GoConfigInvalidException;
 import com.thoughtworks.go.config.exceptions.RecordNotFoundException;
+import com.thoughtworks.go.config.rules.RuleAwarePluginProfile;
+import com.thoughtworks.go.config.rules.RuleAwarePluginProfiles;
 import com.thoughtworks.go.config.update.RuleAwarePluginProfileCommand;
 import com.thoughtworks.go.domain.config.ConfigurationProperty;
 import com.thoughtworks.go.plugin.api.response.validation.ValidationError;
